Update removeWater.py
Had to remove +6 because it was causing the start of the pdf file to be cut off which caused it to not be read by adobe acrobat. The reason i added that in the first place to to try to keep the stream and streamend indicators in the pdf file. these are not necessary however so we just remove them